deeplearning_ai/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
jupyter>=1.0.0
ipykernel>=6.29.0
